+++ anyim-chat/src/main/java/com/hibob/anyim/chat/entity/SessionChat.java
@@ -22,10 +22,10 @@ public class SessionChat {
     private String userB;
 
     @TableField(value = "ref_msg_id")
-    private int refMsgId;
+    private long refMsgId;
 
     @TableField(value = "new_msg_id")
-    private int newMsgId;
+    private long newMsgId;
 
     @TableField(value = "new_time")
     private Date newTime;
